Split the 569-line runtime_diagnostics module into a package (failure_classification + artifact_status) behind a re-exporting __init__ with explicit __all__, plus a surface test — mechanical move, no logic change
